Move code to viewers.py
